numcodecs_ebcc

Function compress

pub fn compress<S: Data<Elem = f32>, D: Dimension>(
    data: ArrayBase<S, D>,
    residual: EbccResidualType,
    base_cr: Positive<f32>,
) -> Result<Vec<u8>, EbccCodecError>

Compress the data array using EBCC with the provided residual and base_cr.

Errors

Errors with

EbccCodecError::HeaderEncodeFailed if encoding the header failed

EbccCodecError::InsufficientDimensions if the data has fewer than two dimensions or the last two dimensions are not at least 32x32

EbccCodecError::EbccEncodeFailed if encoding with EBCC failed

EbccCodecError::SliceEncodeFailed if encoding a 3D slice failed